% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apastyle.R
\docType{data}
\name{apatheme}
\alias{apatheme}
\title{APA style for ggplot2}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 57.}
\usage{
apatheme
}
\description{
implements a ggplot2 style for APA 6th.  Blank bg, no gridlines, bleak.
}
\keyword{datasets}
