% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parPlot.R
\name{parPlot}
\alias{parPlot}
\title{parPlot}
\usage{
parPlot(x)
}
\arguments{
\item{x}{a list of output from the psych::fa.parallel function}
}
\description{
Generates a ggplot of a parallel analysis
}
\examples{
df <- psych::bfi
parallel <- psych::fa.parallel(df, fa= "fa", quant = .95)
parPlot(parallel)
p
ggsave('parallel.png', p, width=6, height=6, unit='in', dpi=300)
}
